pushing changes

+++ Peaksware/PeakswareTest/DTO/IDataChannel.cs
@@ -4,7 +4,7 @@ namespace PeakswareTest.DTO
 {
     public interface IDataChannel
     {
-        Dictionary<int, double> getMaxEfforts();
+        Dictionary<int, int> getMaxEfforts();
         Dictionary<double, ushort> getData();
         void setData(Dictionary<double, ushort> rawData);
         void calculateMaxEffort(int effortTimeMinutes);
