+++ Peaksware/PeakswareTest/Views/ConsoleView.cs
@@ -14,9 +14,9 @@ namespace PeakswareTest.Views
             return inputFile;
         }
 
-        public static void reportEfforts(Dictionary<int, double> efforts)
+        public static void reportEfforts(Dictionary<int, int> efforts)
         {
-            foreach (KeyValuePair<int, double> effort in efforts)
+            foreach (KeyValuePair<int, int> effort in efforts)
             {
                 Console.WriteLine("Best effor for {0} minutes during this ride was: {1} Watts!", effort.Key, effort.Value);
             }
